AC machine selector
This code is meant to help selecting AC machines. It has a feature to size duct types machines and the grills needed in each room. Intially planned for its use in my current company (that is a distributor of AC machines) to help gaining productivity and increase the speed of the service offered.